[Refactor] 상품 조회 메서드 분리
