+++ fridaymarket_resource/src/main/java/com/smile/fridaymarket_resource/domain/product/repository/ProductRepository.java
@@ -1,5 +1,6 @@
-package com.smile.fridaymarket_resource.domain.product.entity;
+package com.smile.fridaymarket_resource.domain.product.repository;
 
+import com.smile.fridaymarket_resource.domain.product.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
