+++ fridaymarket_resource/src/main/java/com/smile/fridaymarket_resource/domain/product/service/ProductReader.java
@@ -1,0 +1,9 @@
+package com.smile.fridaymarket_resource.domain.product.service;
+
+import com.smile.fridaymarket_resource.domain.product.entity.Product;
+
+public interface ProductReader {
+
+    Product getProduct(Long productId);
+
+}
